+++ src/Generators/Tests/Laravel/Controller/ControllerTestGenerator.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpUnitGen\Core\Generators\Tests\Laravel\Controller;
+
+use PhpUnitGen\Core\Contracts\Generators\Factories\ClassFactory as ClassFactoryContract;
+use PhpUnitGen\Core\Contracts\Generators\Factories\MethodFactory as MethodFactoryContract;
+use PhpUnitGen\Core\Generators\Tests\Concerns\ChecksMethods;
+use PhpUnitGen\Core\Generators\Tests\Laravel\FeatureClassFactory;
+use PhpUnitGen\Core\Generators\Tests\Laravel\LaravelTestGenerator;
+use PhpUnitGen\Core\Models\TestClass;
+use Roave\BetterReflection\Reflection\ReflectionMethod;
+
+/**
+ * Class ControllerTestGenerator.
+ *
+ * @author  Paul Thébaud <paul.thebaud29@gmail.com>
+ * @author  Killian Hascoët <killianh@live.fr>
+ * @license MIT
+ */
+class ControllerTestGenerator extends LaravelTestGenerator
+{
+    use ChecksMethods;
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function implementations(): array
+    {
+        return array_merge(parent::implementations(), [
+            ClassFactoryContract::class  => FeatureClassFactory::class,
+            MethodFactoryContract::class => ControllerMethodFactory::class,
+        ]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function isTestable(TestClass $class, ReflectionMethod $reflectionMethod): bool
+    {
+        return $this->config->automaticGeneration();
+    }
+}